% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot a histogram}
\usage{
plot_histogram(data, var, fill_color = "skyblue")
}
\arguments{
\item{data}{A data frame containing the variable to plot.}

\item{var}{The variable to plot (unquoted).}

\item{fill_color}{Color to fill the bars. Default is "skyblue".}
}
\value{
A ggplot object.
}
\description{
This function creates a histogram of a numeric variable.
}
\examples{
plot_histogram(mtcars, mpg)
}
