% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot a scatter plot}
\usage{
plot_scatter(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The name of the variable to plot on the x-axis (unquoted).}

\item{y}{The name of the variable to plot on the y-axis (unquoted).}
}
\value{
A ggplot object.
}
\description{
This function creates a scatter plot between two numeric variables.
}
\examples{
plot_scatter(mtcars, mpg, wt)
}
